updated cloudinary utility for cloud storage of images

+++ pages/api/delete/[id].ts
@@ -1,4 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
+import { DeleteApiResponse, deleteImage } from "@utils/cloudinary";
 import prisma from "@utils/prisma";
 
 /**
@@ -16,10 +17,13 @@ export default async function handle(
 ) {
   const publicId = req.body?.id;
   if (req.method === "DELETE") {
-    const post = await prisma.post.delete({
-      where: { id: publicId },
-    });
-    res.json(post);
+    const response = (await deleteImage(publicId)) as DeleteApiResponse;
+    if (response.message === "ok") {
+      const post = await prisma.post.delete({
+        where: { id: publicId },
+      });
+      res.json(post);
+    }
   } else {
     throw new Error(
       `The HTTP ${req.method} method is not supported at this route.`
